refactored code to use a promise provider rather than depending on the browser field
Also fixed all the inconsistencies with single vs double quotes
throughout the project.
Fixed task runner example
