+++ example/taskrunner/src/taskrunner.js
@@ -1,5 +1,5 @@
 var Task      = require("./task");
-var Bitloader = require("../../dist/bit-loader.js");
+var Bitloader = require("bit-loader");
 var Utils     = Bitloader.Utils;
 var Promise   = Bitloader.Promise;
