+++ src/module/linker.js
@@ -1,7 +1,7 @@
 (function(root) {
   "use strict";
 
-  var logger = require('logger').factory("Module/Linker");
+  var logger = require("../logger").factory("Module/Linker");
 
   function ModuleLinker(manager, mod) {
     function traverseDependencies(mod) {
@@ -27,4 +27,4 @@
   }
 
   module.exports = ModuleLinker;
-})(typeof(window) !== 'undefined' ? window : this);
+})(typeof(window) !== "undefined" ? window : this);
