+++ src/promise.js
@@ -1,0 +1,4 @@
+(function() {
+  "use strict";
+  module.exports = require("spromise");
+})();
